Remove effective area as function of reco energy, fixes #137
This removes all mentions of effective area as a function of
reconstructed energy as this is not possible to calculate.

The assumption of unity energy migration can of course still be done,
but that changes nothing concerning the fact that effective area can
only ever be expressed in bins of true, simulated energy.